Fix volume increase when channel names have digits
On my PC, the channel names for left/right speaker on one of my sinks
are named aux0 and aux1. Because of the particular construction of the
regex in `pulseaudio-control-increase-volume`, the volume string was not
parsed correctly. This commit corrects this by searching for a ":"
rather than searching for the next non-digit when parsing the second
channel's volume information.